Add JWT authentication with signup, login, and refresh token support

+++ airBnbApp/src/main/java/com/maskara/airBnbApp/service/HotelServiceImpl.java
@@ -4,14 +4,17 @@ import com.maskara.airBnbApp.dto.HotelDto;
 import com.maskara.airBnbApp.dto.HotelInfoDto;
 import com.maskara.airBnbApp.dto.RoomDto;
 import com.maskara.airBnbApp.exception.ResourceNotFoundException;
+import com.maskara.airBnbApp.exception.UnAuthorisedException;
 import com.maskara.airBnbApp.modal.Hotel;
 import com.maskara.airBnbApp.modal.Room;
+import com.maskara.airBnbApp.modal.User;
 import com.maskara.airBnbApp.repository.HotelRepository;
 import com.maskara.airBnbApp.repository.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,6 +34,10 @@ public class HotelServiceImpl implements HotelService{
         log.info("Creating new Hotel with name: {}",hotelDto.getName());
         Hotel hotel = modelMapper.map(hotelDto,Hotel.class);
         hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel = hotelRepository.save(hotel);
         log.info("Created new Hotel with id: {}",hotelDto.getId());
         return modelMapper.map(hotel, HotelDto.class);
@@ -42,6 +49,11 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with id: "+id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
         return modelMapper.map(hotel, HotelDto.class);
     }
 
@@ -51,6 +63,13 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with id: "+id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         modelMapper.map(hotelDto, hotel);
         hotel.setId(id);
         hotel=hotelRepository.save(hotel);
@@ -63,7 +82,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with Id: "+id));
-        hotelRepository.deleteById(id);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+id);
+        }
+//        hotelRepository.deleteById(id);
         for(Room room: hotel.getRooms()){
             inventoryService.deleteFutureInventory(room);
             roomRepository.deleteById(room.getId());
@@ -82,6 +106,11 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with id: "+hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())){
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
         hotel.setActive(true);
 
         for(Room room: hotel.getRooms()){
@@ -91,7 +120,7 @@ public class HotelServiceImpl implements HotelService{
 
 
     }
-
+//   public method
     @Override
     public HotelInfoDto getHotelInfoById(Long hotelId) {
         Hotel hotel = hotelRepository
